Cardano.Crypto.KES.ForgetMock

Description

Mock key evolving signatures.

Documentation

data ForgetMockKES k

A wrapper for a KES implementation that adds logging functionality, for the purpose of verifying that invocations of genKeyKES and forgetSignKeyKES pair up properly in a given host application.

The wrapped KES behaves exactly like its unwrapped payload, except that invocations of genKeyKES, updateKES and forgetSignKeyKES are logged as GenericEvent ForgetMockEvent values. (We use GenericEvent in order to use the generic MonadSodium instance of LogT; otherwise we would have to provide a boilerplate instance here).

data family VerKeyKES v

data family SignKeyKES v

data family SigKES v

data ForgetMockEvent

Constructors

GEN Word

UPD Word Word

NOUPD

DEL Word

isGEN :: ForgetMockEvent -> Bool

isUPD :: ForgetMockEvent -> Bool

isDEL :: ForgetMockEvent -> Bool