Test.Crypto.AllocLog

Documentation

data AllocEvent

Allocation log event. These are emitted automatically whenever mlocked memory is allocated through the mlockedAllocForeignPtr primitive, or released through an associated finalizer (either explicitly or due to GC). Additional events that are not actual allocations/deallocations, but may provide useful debugging context, can be inserted as MarkerEv.

Constructors

AllocEv !WordPtr

FreeEv !WordPtr

MarkerEv !String

newtype LogT event m a

Constructors

LogT

Fields

unLogT :: ReaderT (Tracer (LogT event m) event) m a

type AllocLogT = LogT AllocEvent

runLogT :: Tracer (LogT event m) event -> LogT event m a -> m a

runAllocLogT :: Tracer (LogT AllocEvent m) AllocEvent -> LogT AllocEvent m a -> m a

pushLogEvent :: Monad m => event -> LogT event m ()

pushAllocLogEvent :: Monad m => AllocEvent -> LogT AllocEvent m ()

newtype GenericEvent e

Newtype wrapper over an arbitrary event; we use this to write the generic MonadSodium instance below while avoiding overlapping instances.

Constructors

GenericEvent

Fields

concreteEvent :: e