Test.Crypto.EqST

Documentation

class EqST a where

Monadic flavor of Eq, for things that can only be compared in a monadic context that satisfies MonadST. This is needed because we cannot have a sound Eq instance on mlocked memory types, but we do need to compare them for equality in tests.

Methods

equalsM ∷ MonadST m ⇒ a → a → m Bool

nequalsM ∷ (MonadST m, EqST a) ⇒ a → a → m Bool

(==!) ∷ (MonadST m, EqST a) ⇒ a → a → m Bool infix 4

Infix version of equalsM

(!=!) ∷ (MonadST m, EqST a) ⇒ a → a → m Bool infix 4

Infix version of nequalsM

newtype PureEqST a

Helper newtype, useful for defining EqST in terms of Eq for types that have sound Eq instances, using DerivingVia. An Applicative context must be provided for such instances to work, so this will generally require StandaloneDeriving as well.

Ex.: deriving via PureEq Int instance Applicative m => EqST m Int

Constructors

PureEqST a