Test.Crypto.Util

CBOR

class Typeable a ⇒ FromCBOR a where

Methods

fromCBOR ∷ Decoder s a

class Typeable a ⇒ ToCBOR a where

Minimal complete definition

toCBOR

Methods

toCBOR ∷ a → Encoding

encodedSizeExpr ∷ (∀ t. ToCBOR t ⇒ Proxy t → Size) → Proxy a → Size

prop_cbor ∷ (ToCBOR a, FromCBOR a, Eq a, Show a) ⇒ a → Property

prop_cbor_size ∷ ∀ a. ToCBOR a ⇒ a → Property

prop_cbor_with ∷ (Eq a, Show a) ⇒ (a → Encoding) → (∀ s. Decoder s a) → a → Property

prop_cbor_valid ∷ (a → Encoding) → a → Property

prop_cbor_roundtrip ∷ (Eq a, Show a) ⇒ (a → Encoding) → (∀ s. Decoder s a) → a → Property

prop_raw_serialise ∷ (Eq a, Show a) ⇒ (a → ByteString) → (ByteString → Maybe a) → a → Property

prop_raw_deserialise ∷ ∀ (a ∷ Type). Show a ⇒ (ByteString → Maybe a) → BadInputFor a → Property

prop_size_serialise ∷ (a → ByteString) → Word → a → Property

prop_cbor_direct_vs_class ∷ ToCBOR a ⇒ (a → Encoding) → a → Property

The crypto algorithm classes have direct encoding functions, and the key types are also typically a member of the ToCBOR class. Where a ToCBOR instance is provided then these should match.

NoThunks

prop_no_thunks ∷ NoThunks a ⇒ a → Property

When forcing the given value to WHNF, it may no longer contain thunks.

prop_no_thunks_IO ∷ NoThunks a ⇒ IO a → IO Property

prop_no_thunks_IO_from ∷ NoThunks a ⇒ (b → IO a) → b → Property

prop_no_thunks_IO_with ∷ NoThunks a ⇒ Gen (IO a) → Property

Test Seed

newtype TestSeed

Constructors

TestSeed

Fields

getTestSeed ∷ (Word64, Word64, Word64, Word64, Word64)

withTestSeed ∷ TestSeed → MonadPseudoRandom ChaChaDRG a → a

testSeedToChaCha ∷ TestSeed → ChaChaDRG

nullTestSeed ∷ TestSeed

Seeds

data SizedSeed (n ∷ Nat)

unSizedSeed ∷ SizedSeed n → Seed

arbitrarySeedOfSize ∷ Word → Gen Seed

arbitrarySeedBytesOfSize ∷ Word → Gen ByteString

test messages for signings

newtype Message

Constructors

Message

Fields

messageBytes ∷ ByteString

Test generation and shrinker helpers

data BadInputFor (a ∷ Type)

genBadInputFor ∷ ∀ (a ∷ Type). Int → Gen (BadInputFor a)

shrinkBadInputFor ∷ ∀ (a ∷ Type). BadInputFor a → [BadInputFor a]

showBadInputFor ∷ ∀ (a ∷ Type). BadInputFor a → String

Formatting

hexBS ∷ ByteString → String

Helpers for testing IO actions

noExceptionsThrown ∷ Applicative m ⇒ m Property

Return a property that always succeeds in some monad (typically IO). This is useful to express that we are only interested in whether the side effects of the preceding actions caused any exceptions or not - if they did, then the test will fail because of it, but if they did not, then noExceptionsThrown will be reached, and the test will succeed.

doesNotThrow ∷ Applicative m ⇒ m a → m Property

Chain monadic action with noExceptionsThrown to express that we only want to make sure that the action does not throw any exceptions, but we are not interested in its result.

Direct ser/deser helpers

directSerialiseToBS ∷ ∀ m a. DirectSerialise a ⇒ MonadST m ⇒ MonadThrow m ⇒ Int → a → m ByteString

directDeserialiseFromBS ∷ ∀ m a. DirectDeserialise a ⇒ MonadST m ⇒ MonadThrow m ⇒ ByteString → m a

Error handling

eitherShowError ∷ (HasCallStack, Show e) ⇒ Either e a → IO a

Locking

data Lock

withLock ∷ Lock → IO a → IO a

mkLock ∷ IO Lock