Test.Crypto.AllocLog

Documentation

data AllocEvent

Allocation log event. These are emitted automatically whenever mlocked memory is allocated through the mlockedAllocForeignPtr primitive, or released through an associated finalizer (either explicitly or due to GC). Additional events that are not actual allocations/deallocations, but may provide useful debugging context, can be inserted as MarkerEv.

Constructors

AllocEv !WordPtr

FreeEv !WordPtr

MarkerEv !String

mkLoggingAllocator ∷ Tracer IO AllocEvent → MLockedAllocator IO → MLockedAllocator IO